requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.5
dlib==19.24.0
docopt==0.6.2
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
filelock==3.8.0
Flask==2.1.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.3
imageio==2.22.4
importlib-metadata==4.12.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
networkx==2.8.8
numpy==1.23.2
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.2.0
pipenv==2022.8.24
pipreqs==0.4.11
platformdirs==2.5.2
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
tifffile==2022.10.10
urllib3==1.26.12
virtualenv==20.16.3
virtualenv-clone==0.5.7
Wand==0.6.10
Werkzeug==2.1.2
yarg==0.1.9
zipp==3.8.0
